Using CAS operations instead of synchronized in ValidationSupportChain
This change replaces the handling of cached StructureDefinitions in
ValidationSupportChain by the use of a new ResourceAggregator class.
This class implements a copy-on-write strategy for cached resources
and makes it easier to ensure thread safety by using (de-facto)
immutable state and an AtomicReference to swap out the state in a
single atomic operation.